Module reth_rpc_types_compat::block

Compatibility functions for rpc Block type.

Functions

from_block

Converts the given primitive block into a Block response with the given BlockTransactionsKind

from_block_full

Create a new Block response from a primitive block, using the total difficulty to populate its field in the rpc response.

from_block_with_tx_hashes

Create a new Block response from a primitive block, using the total difficulty to populate its field in the rpc response.

from_primitive_with_hash

Converts from a reth_primitives::SealedHeader to a reth_rpc_types::BlockNumberOrTag

uncle_block_from_header

Build an RPC block response representing an Uncle from its header.