Function reth_rpc_types_compat::block::from_block

pub fn from_block(
    block: BlockWithSenders,
    total_difficulty: U256,
    kind: BlockTransactionsKind,
    block_hash: Option<B256>
) -> Result<Block, BlockError>

Converts the given primitive block into a Block response with the given BlockTransactionsKind

If a block_hash is provided, then this is used, otherwise the block hash is computed.