Module reth_rpc_types_compat::engine::payload

Standalone Conversion Functions for Handling Different Versions of Execution Payloads in Ethereum’s Engine

Functions

block_to_payload_v3

Converts SealedBlock to ExecutionPayloadV3

convert_block_to_payload_field_v2

Converts SealedBlock to ExecutionPayloadFieldV2

convert_block_to_payload_input_v2

Converts SealedBlock to ExecutionPayloadInputV2

convert_payload_field_v2_to_payload

Converts ExecutionPayloadFieldV2 to ExecutionPayload

convert_payload_input_v2_to_payload

Converts ExecutionPayloadInputV2 to ExecutionPayload

convert_standalone_withdraw_to_withdrawal

Converts reth_rpc_types::engine::payload::Withdrawal to Withdrawal

convert_to_payload_body_v1

Converts Block to ExecutionPayloadBodyV1

convert_withdrawal_to_standalone_withdraw

Converts Withdrawal to reth_rpc_types::engine::payload::Withdrawal

execution_payload_from_sealed_block

Transforms a SealedBlock into a ExecutionPayloadV1

try_block_to_payload

Converts SealedBlock to ExecutionPayload

try_block_to_payload_v1

Converts SealedBlock to ExecutionPayloadV1

try_block_to_payload_v2

Converts SealedBlock to ExecutionPayloadV2

try_into_block

Tries to create a new block (without a block hash) from the given payload and optional parent beacon block root. Performs additional validation of extra_data and base_fee_per_gas fields.

try_into_sealed_block

Tries to create a new block from the given payload and optional parent beacon block root.

try_payload_v1_to_block

Converts ExecutionPayloadV1 to Block

try_payload_v2_to_block

Converts ExecutionPayloadV2 to Block

try_payload_v3_to_block

Converts ExecutionPayloadV3 to Block

validate_block_hash

Takes the expected block hash and Block, validating the block and converting it into a SealedBlock.