Function reth_rpc_types_compat::engine::payload::try_into_block

pub fn try_into_block(
    value: ExecutionPayload,
    parent_beacon_block_root: Option<B256>
) -> Result<Block, PayloadError>

Tries to create a new block (without a block hash) from the given payload and optional parent beacon block root. Performs additional validation of extra_data and base_fee_per_gas fields.

NOTE: The log bloom is assumed to be validated during serialization.

See https://github.com/ethereum/go-ethereum/blob/79a478bb6176425c2400e949890e668a3d9a3d05/core/beacon/types.go#L145